patch-one
includes
- social login images
- bank modal and tranfer model on dashboard log
- touch up on the transaction log
