+++ src/components/transaction-log/transaction-log.component.html
@@ -1,6 +1,6 @@
 
 <div class="transaction-list">
   <ng-container *ngFor="let transaction of transactions">
-    <transaction [transaction]="transaction" (click)="showTransferModal(transaction)"></transaction>
+    <transaction [transaction]="transaction"></transaction>
   </ng-container>
 </div>
