+++ src/components/transaction-log/transaction-log.component.ts
@@ -1,7 +1,5 @@
 import { Component, Input } from "@angular/core";
 import { HomePage } from "../../pages/home/home";
-import { TransactionsLogModalPage } from "../../pages/modals/transactionslog-modal/transactionslog-modal";
-import { ModalController } from "ionic-angular";
 
 @Component({
   selector: "transaction-log",
@@ -14,7 +12,6 @@ export class TransactionLogComponent {
 
   constructor(
     public home: HomePage,
-    public modalCtrl: ModalController,
   ) {
     this.transactions = this.getWallet()[0].transactions.data;
     this.transactions = this.transactions.reverse();
@@ -34,15 +31,5 @@ export class TransactionLogComponent {
    */
   public getWallet() {
     return this.home.wallets;
-  }
-
-  /**
-   * Displays the modal page for the given transfer.
-   *
-   * @param transfer
-   */
-  public showTransferModal(transfer) {
-    let modal = this.modalCtrl.create(TransactionsLogModalPage, { transfer });
-    modal.present();
   }
 }
