+++ src/components/transaction/transaction.component.ts
@@ -1,4 +1,9 @@
 import { Component, Input } from "@angular/core";
+import { ModalController } from "ionic-angular";
+import { AuthHttp } from "angular2-jwt";
+import { environment } from "../../environments/environment";
+import { BankTransferModalPage } from "../../pages/modals/banktransfer-modal/banktransfer-modal";
+import {TransfersModalPage} from "../../pages/modals/transfers-modal/transfers-modal";
 
 @Component({
   selector: "transaction",
@@ -7,5 +12,29 @@ import { Component, Input } from "@angular/core";
 export class TransactionComponent {
 
   @Input() public transaction: any;
+  constructor(
+    private modalCtrl: ModalController,
+    private http: AuthHttp,
+  ) {}
 
+  /**
+   * Displays the modal page for the given transfer.
+   *
+   * @param transfer
+   */
+  public showTransferModal(transaction) {
+    if (transaction.transaction_type === "bank") {
+      this.http.get(environment.server_url + "/api/transfer/bank/transfer/" + transaction.transaction_id)
+        .map((res) => res.json())
+        .subscribe((res) => {
+          this.modalCtrl.create(BankTransferModalPage, { transfer: res.data }).present();
+        });
+    } else {
+      this.http.get(environment.server_url + "/api/transfer/user/transfer/" + transaction.transaction_id)
+        .map((res) => res.json())
+        .subscribe((res) => {
+          this.modalCtrl.create(TransfersModalPage, { transfer: res.data }).present();
+        });
+    }
+  }
 }
