--- src/models/User.ts
@@ -1,8 +1,0 @@
-export class User {
-    constructor(
-      public id: number,
-      public name: string,
-      public phoneNumber: string,
-      public email: string,
-    ) { }
-}
