+++ src/pages/login/login.html
@@ -1,35 +1,13 @@
 
-<ion-content padding [fullscreen]="true"
-			 style="background: url('assets/images/login-background.jpg') no-repeat center; background-size: cover">
+<ion-content padding [fullscreen]="true" style="background: url('assets/images/login-background.jpg') no-repeat center; background-size: cover">
 	<div class="logo">
 		<ion-img src="assets/images/logo.png" [width]="160" [height]="85"></ion-img>
 	</div>
-	<!--<form (onSubmit)="login(form)" #form="ngForm">-->
-		<!--<div class="input-group">-->
-			<!--<ion-icon name="person"></ion-icon>-->
-			<!--<ion-input-->
-					<!--type="email"-->
-					<!--placeholder="Email"-->
-					<!--required-->
-			<!--&gt;</ion-input>-->
-		<!--</div>-->
-		<!--<div class="input-group">-->
-			<!--<ion-icon name="lock"></ion-icon>-->
-			<!--<ion-input-->
-					<!--type="text"-->
-					<!--placeholder="Password"-->
-			<!--&gt;</ion-input>-->
-		<!--</div>-->
-		<!--<button ion-button block type="submit">Sign In</button>-->
-	<!--</form>-->
-	<!--<div class="seperator">-->
-		<!--or sign in with-->
-	<!--</div>-->
-  <button class="btn-signin" ion-button block icon-left (click)="showAuth('facebook')">
-    Facebook
+  <button class="social-login facebook-login" ion-button block (click)="showAuth('facebook')">
+      <img src="assets/images/facebook-logo.svg" alt=""><div>Log in with Facebook</div>
   </button>
-  <button class="btn-signin" ion-button block icon-left (click)="showAuth('google')">
-    Google
+  <button class="social-login google-login" ion-button block (click)="showAuth('google')">
+      <img src="assets/images/google-logo.svg" alt=""> <div>Sign in with Google</div>
   </button>
 
   <button style="opacity: 0.01;" class="btn-signin" ion-button block icon-left (click)="autoLogin()">
