+++ src/pages/modals/banktransfer-modal/banktransfer-modal.scss
@@ -1,0 +1,3 @@
+page-banktransfer-modal {
+
+}
