--- src/pages/modals/transactionslog-modal/transactionslog-modal.scss
@@ -1,4 +1,0 @@
-/* tslint:disable:no-empty */
-page-transactionslog-modal {
-
-}
