+++ src/pages/modals/transfers-modal/transfers-modal.ts
@@ -1,5 +1,5 @@
 import { Component } from "@angular/core";
-import { ViewController, NavParams } from "ionic-angular";
+import { NavParams, ViewController } from "ionic-angular";
 
 @Component({
   selector: "page-transfers-modal",
@@ -8,13 +8,15 @@ import { ViewController, NavParams } from "ionic-angular";
 export class TransfersModalPage {
 
   public transfer;
-
   constructor(
     public viewCtrl: ViewController,
     public params: NavParams) {
     this.transfer = this.params.get("transfer");
   }
 
+  /**
+   * Dismiss view
+   */
   public dismiss() {
     this.viewCtrl.dismiss();
   }
